requirement.txt
streamlit
pandas
langchain-google-genai
python-decouple
langchain
langchain-community
